requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attmap==0.13.2
attrs==24.2.0
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cmudict==1.0.30
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.6
ctranslate2==4.3.1
Cython==3.0.11
dataclasses-json==0.6.7
Deprecated==1.2.14
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
eido==0.2.2
emoji==2.12.1
exceptiongroup==1.2.2
fastapi==0.112.2
faster-whisper==1.0.3
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
g2p-en==2.1.0
google-auth==2.34.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
inflect==7.3.1
intel-openmp==2021.4.0
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.15
langchain-chroma==0.1.3
langchain-community==0.2.15
langchain-core==0.2.37
langchain-text-splitters==0.2.2
langsmith==0.1.108
llvmlite==0.43.0
logmuse==0.2.8
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.2
onnxruntime==1.19.0
openai==1.43.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
oyaml==1.0
packaging==24.1
pandas==2.2.2
pephubclient==0.4.4
peppy==0.40.5
pillow==10.2.0
pipestat==0.10.2
posthog==3.6.0
proces==0.1.7
protobuf==4.25.4
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyopenjtalk-dict==0.3.4.dev2
pypdf==4.3.1
PyPika==0.48.9
pypinyin==0.52.0
pypiwin32==223
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pytchat==0.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.91
pytz==2024.1
pywin32==306
pyworld-prebuilt==0.3.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.32
starlette==0.38.4
style-bert-vits2==2.5.0
sympy==1.13.2
tbb==2021.11.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.5
transformers==4.44.2
typeguard==4.3.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ubiquerg==0.8.0
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.0.1
win32-setctime==1.1.0
wrapt==1.16.0
yacman==0.9.3
yarl==1.9.7
zipp==3.20.1
